requirements.txt
numpy
scipy
Pillow
pandas
matplotlib
scikit-learn
tensorflow>=2.9.0
keras>=2.9.0
opencv-python
